+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release notes
 ==============
 
+### Unreleased
+
+* Fixed incorrect subject name when `.Should` was placed right after `WhoseValue`
+
 ### 1.3.3 (2024-01-29)
 
 * Updated YamlDotNet from 13.3.1 to 15.1.0